requirements.txt
aiohappyeyeballs>=2.4.6
aiohttp>=3.11.12
aiosignal>=1.3.2
attrs>=25.1.0
beautifulsoup4>=4.13.3
Brotli>=1.1.0
bs4>=0.0.2
certifi>=2025.1.31
cffi>=1.17.1
charset-normalizer>=3.4.1
discord.py>=2.5.2
dnspython>=2.7.0
frozenlist>=1.5.0
idna>=3.10
multidict>=6.1.0
mutagen>=1.47.0
propcache>=0.2.1
pycparser>=2.22
pycryptodomex>=3.21.0
pymongo>=4.11.1
PyNaCl>=1.5.0
python-dotenv>=1.0.1
pytube>=15.0.0
requests>=2.32.3
soupsieve>=2.6
typing_extensions>=4.12.2
urllib3>=2.3.0
websockets>=14.2
yarl>=1.18.3
yt-dlp>=2025.1.26
